Refactor scripts into package and update docs
