+++ copernican_lib/__init__.py
@@ -1,0 +1,7 @@
+"""Core helper modules for the Copernican Suite.
+
+This package bundles general-purpose utilities used across the
+`copernican` command-line tool and the computational engines.  Modules
+here implement data loading, plotting and other shared logic so that
+individual engines remain lightweight.
+"""
