+++ copernican_lib/model_coder.py
