+++ copernican_lib/model_parser.py
