+++ data/gw/placeholder/cosmo_parser_gw_placeholder.py
@@ -1,6 +1,6 @@
 import logging
 import os
-from scripts.data_loaders import register_gw_parser
+from copernican_lib.data_loaders import register_gw_parser
 
 @register_gw_parser("gw_placeholder_v1", "Placeholder GW parser.", data_dir=os.path.dirname(__file__))
 def parse_gw_placeholder(data_dir, **kwargs):
