+++ engines/cosmo_engine_comb.py
@@ -4,7 +4,7 @@
 This backend performs a simultaneous fit of cosmological parameters to all
 available datasets (SNe Ia, BAO and CMB).  Numerical functions for distances
 and the sound horizon are provided by the model plugin generated by
-:mod:`scripts.model_coder`.  The engine itself only orchestrates the
+:mod:`copernican_lib.model_coder`.  The engine itself only orchestrates the
 optimisation and calculation of chi-squared values.
 """
 
@@ -14,8 +14,8 @@ import camb
 from functools import lru_cache
 import sys
 import logging
-from scripts import engine_interface
-from scripts.optim_utils import minimize_with_progress
+from copernican_lib import engine_interface
+from copernican_lib.optim_utils import minimize_with_progress
 
 # --- Constants for SNe H2-style (SALT2 nuisance parameter fitting) ---
 SALT2_NUISANCE_PARAMS_INIT = {
