--- scripts/error_handler.py
@@ -1,7 +1,0 @@
-"""Simple error logging utility for the JSON pipeline."""
-
-import logging
-
-
-def report_error(message):
-    logging.getLogger().error(message)
